Update comments with reasoning in ptr_guaranteed_cmp.
`GlobalAlloc::TypeId` exists mostly to prevent consteval from comparing `TypeId`s, so always
return 2 for those (whether the pointers are to the same allocation or not).
In the different-allocation case, document that `GlobalAlloc::Memory`, `Function`, and `Vtable` can
be deduplicated, at least with other allocations of the same kind, so those should return 2.