Work towards allowing non-`usize`-sized pointer metadata in codegen.
Adds a typestate for `rustc_codegen_ssa::OperandRef/OperandValue/PlaceRef/PlaceValue` for if the operand/place can be unsized or not,
and use (a wrapper around) `Option<OperandRef<SizedPlace>>` as the pointer metadata in `PlaceValue<AnyPlace>`.

Currently, everything else stays with the existing behavior of "could be unsized" (`AnyPlace`), but we could explore restricting some
things to "must be a sized place" at the type level in some places instead of having `assert!(!llextra.has_metadata())`.

[FIXME: rework this?] PlaceRef::project_fields needs to be able to `OperandRef::extract_field` to get metadata without an alloca,
but `extract_field` needs a `FunctionCx`, so I made a copy of it that does fewer checks (i.e. just returns self.val instead of
fx.codegen_transmute_operand(..) if the field('s metadata) is the same size as the container('s metadata)).